Removed Attribute from Model name since namespace already is Attribute
