+++ lib/fog/core.rb
@@ -13,14 +13,14 @@ require 'ipaddr'
 # internal core dependencies
 require "fog/version"
 require 'fog/core/attributes'
-require 'fog/core/attributes/default_attribute'
-require 'fog/core/attributes/boolean_attribute'
-require 'fog/core/attributes/float_attribute'
-require 'fog/core/attributes/integer_attribute'
-require 'fog/core/attributes/string_attribute'
-require 'fog/core/attributes/time_attribute'
-require 'fog/core/attributes/timestamp_attribute'
-require 'fog/core/attributes/array_attribute'
+require 'fog/core/attributes/default'
+require 'fog/core/attributes/boolean'
+require 'fog/core/attributes/float'
+require 'fog/core/attributes/integer'
+require 'fog/core/attributes/string'
+require 'fog/core/attributes/time'
+require 'fog/core/attributes/timestamp'
+require 'fog/core/attributes/array'
 require 'fog/core/collection'
 require 'fog/core/connection'
 require 'fog/core/credentials'
