+++ lib/fog/core/attributes/float.rb
@@ -1,10 +1,10 @@
 module Fog
   module Attributes
-    class IntegerAttribute < DefaultAttribute
+    class Float < Default
       def create_setter
         model.class_eval <<-EOS, __FILE__, __LINE__
             def #{name}=(new_#{name})
-              attributes[:#{name}] = new_#{name}.to_i
+              attributes[:#{name}] = new_#{name}.to_f
             end
         EOS
       end
