+++ lib/fog/core/attributes/integer.rb
@@ -1,11 +1,11 @@
 module Fog
   module Attributes
-    class ArrayAttribute < DefaultAttribute
+    class Integer < Default
       def create_setter
         model.class_eval <<-EOS, __FILE__, __LINE__
-          def #{name}=(new_#{name})
-            attributes[:#{name}] = [*new_#{name}]
-          end
+            def #{name}=(new_#{name})
+              attributes[:#{name}] = new_#{name}.to_i
+            end
         EOS
       end
     end
