+++ lib/fog/core/attributes/string.rb
@@ -1,10 +1,10 @@
 module Fog
   module Attributes
-    class FloatAttribute < DefaultAttribute
+    class String < Default
       def create_setter
         model.class_eval <<-EOS, __FILE__, __LINE__
             def #{name}=(new_#{name})
-              attributes[:#{name}] = new_#{name}.to_f
+              attributes[:#{name}] = new_#{name}.to_s
             end
         EOS
       end
