+++ lib/fog/core/attributes/time.rb
@@ -1,15 +1,15 @@
 module Fog
   module Attributes
-    class TimeAttribute < DefaultAttribute
+    class Time < Default
       def create_setter
         model.class_eval <<-EOS, __FILE__, __LINE__
             def #{name}=(new_#{name})
-              attributes[:#{name}] = if new_#{name}.nil? || new_#{name} == "" || new_#{name}.is_a?(Time)
+              attributes[:#{name}] = if new_#{name}.nil? || new_#{name} == "" || new_#{name}.is_a?(::Time)
                 new_#{name}
               elsif new_#{name}.respond_to?(:to_time)
                 new_#{name}.to_time
               else
-                Time.parse(new_#{name})
+                ::Time.parse(new_#{name})
               end
             end
         EOS
