+++ lib/fog/core/attributes/timestamp.rb
@@ -1,13 +1,13 @@
 module Fog
   module Attributes
-    class TimestampAttribute < DefaultAttribute
+    class Timestamp < Default
       def create_setter
         model.class_eval <<-EOS, __FILE__, __LINE__
             def #{name}=(new_#{name})
               if new_#{name}.respond_to?(:to_i)
-                attributes[:#{name}] = Time.at(new_#{name}.to_i)
+                attributes[:#{name}] = Fog::Time.at(new_#{name}.to_i)
               else
-                attributes[:#{name}] = Time.parse(new_#{name}.to_s)
+                attributes[:#{name}] = Fog::Time.parse(new_#{name}.to_s)
               end
             end
         EOS
